feat: add reactions functionality to message handling
- Implemented reactions collection interface in docdb with Upsert, Get, ListByMessage, Delete, and DeleteByConversation methods.
- Created MessageReaction model to represent user reactions with validation for reaction types.
- Developed MongoDB implementation for reactions collection, including necessary indexes.
- Added client helper functions for making HTTP requests with bearer and service key authentication.
- Created unit tests for message edit/delete handlers to ensure proper functionality and error handling.
- Added unit tests for reaction handlers, covering upsert, delete, and retrieval of reactions.
