+++ tests/mocks/docdb_mock.go
@@ -112,6 +112,7 @@ type MockDocDBClient struct {
 	mock.Mock
 	messagesCollection    *MockMessagesCollection
 	messagesRawCollection *MockCollection
+	reactionsCollection   *MockReactionsCollection
 	tracesCollection      *MockTracesCollection
 	tracesRawCollection   *MockCollection
 	database              *MockDatabase
@@ -122,6 +123,7 @@ func NewMockDocDBClient() *MockDocDBClient {
 	return &MockDocDBClient{
 		messagesCollection:    &MockMessagesCollection{},
 		messagesRawCollection: &MockCollection{},
+		reactionsCollection:   &MockReactionsCollection{},
 		tracesCollection:      &MockTracesCollection{},
 		tracesRawCollection:   &MockCollection{},
 		database:              &MockDatabase{},
@@ -141,6 +143,11 @@ func (m *MockDocDBClient) Messages() docdb.MessagesCollection {
 // MessagesRaw returns the raw messages collection.
 func (m *MockDocDBClient) MessagesRaw() docdb.Collection {
 	return m.messagesRawCollection
+}
+
+// Reactions returns the typed reactions collection.
+func (m *MockDocDBClient) Reactions() docdb.ReactionsCollection {
+	return m.reactionsCollection
 }
 
 // Traces returns the typed traces collection.
@@ -184,6 +191,11 @@ func (m *MockDocDBClient) GetTracesCollection() *MockTracesCollection {
 // GetTracesRawCollection returns the mock raw traces collection for setup.
 func (m *MockDocDBClient) GetTracesRawCollection() *MockCollection {
 	return m.tracesRawCollection
+}
+
+// GetReactionsCollection returns the mock reactions collection for setup.
+func (m *MockDocDBClient) GetReactionsCollection() *MockReactionsCollection {
+	return m.reactionsCollection
 }
 
 // MockTracesCollection is a mock implementation of docdb.TracesCollection.
@@ -427,6 +439,53 @@ func (m *MockCursor) Err() error {
 
 // Close closes the cursor.
 func (m *MockCursor) Close(ctx context.Context) error {
+	args := m.Called(ctx)
+	return args.Error(0)
+}
+
+// MockReactionsCollection is a mock implementation of docdb.ReactionsCollection.
+type MockReactionsCollection struct {
+	mock.Mock
+}
+
+// Upsert creates or updates a reaction.
+func (m *MockReactionsCollection) Upsert(ctx context.Context, reaction *models.MessageReaction) error {
+	args := m.Called(ctx, reaction)
+	return args.Error(0)
+}
+
+// Get retrieves a reaction.
+func (m *MockReactionsCollection) Get(ctx context.Context, opts *docdb.UpsertReactionOptions) (*models.MessageReaction, error) {
+	args := m.Called(ctx, opts)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).(*models.MessageReaction), args.Error(1)
+}
+
+// ListByMessage retrieves all reactions for a message.
+func (m *MockReactionsCollection) ListByMessage(ctx context.Context, opts *docdb.ListReactionsOptions) ([]*models.MessageReaction, error) {
+	args := m.Called(ctx, opts)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).([]*models.MessageReaction), args.Error(1)
+}
+
+// Delete removes a reaction.
+func (m *MockReactionsCollection) Delete(ctx context.Context, opts *docdb.DeleteReactionOptions) error {
+	args := m.Called(ctx, opts)
+	return args.Error(0)
+}
+
+// DeleteByConversation removes all reactions in a conversation.
+func (m *MockReactionsCollection) DeleteByConversation(ctx context.Context, tenantID, conversationID string) error {
+	args := m.Called(ctx, tenantID, conversationID)
+	return args.Error(0)
+}
+
+// EnsureIndexes creates indexes.
+func (m *MockReactionsCollection) EnsureIndexes(ctx context.Context) error {
 	args := m.Called(ctx)
 	return args.Error(0)
 }
